pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.80.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-docling>=2.0.0",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langsmith>=0.7.20",
    "marimo[recommended]>=0.21.0",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.3.0",
]
